feat: create directory for browser and mobile, separate the files
