+++ pages/prefectures.tsx
@@ -1,11 +1,14 @@
 /* eslint-disable react-hooks/rules-of-hooks */
-import { useRouter } from "next/router";
+// import { useRouter } from "next/router";
 import Link from "next/link";
 import Image from "next/image";
-import { useEffect, useRef } from "react";
+// import { useEffect, useRef } from "react";
+import { PrefList as MobilePrefList } from "./mobile/PrefList";
+import { PrefList as BrowserPrefList } from "./desktop/PrefList";
 import { Container, Box } from "@mui/material";
 import { BrowserView, MobileView } from "react-device-detect";
 import styles from "../styles/prefectures.module.css";
+import { light } from "@mui/material/styles/createPalette";
 
 const engPrefNames = {
   北海道: "Hokkaido",
@@ -57,99 +60,101 @@ const engPrefNames = {
   沖縄県: "Okinawa",
 };
 
+const engPrefArr = Object.values(engPrefNames);
+
 const prefectures = () => {
-  const router = useRouter();
-  const loadingLog = useRef(false);
+  // const router = useRouter();
+  // const loadingLog = useRef(false);
 
-  useEffect(() => {
-    if (loadingLog.current === false) {
-      loadingLog.current = true;
+  // useEffect(() => {
+  //   if (loadingLog.current === false) {
+  //     loadingLog.current = true;
 
-      window.svgJapan({
-        element: "#my-map-container",
-        uniformly: true,
-        uniformColor: "#00984F",
-        // width: "auto",
-        height: "calc(100vh - 120px)",
-        regionality: true,
-        type: "deform",
-        width: "100vw",
-        regions: [
-          {
-            id: 1,
-            name: "Hokkaido",
-            prefs: [1],
-            color: "#689ECA",
-            active: "#F3FF15",
-          },
-          {
-            id: 2,
-            name: "Tohoku",
-            prefs: [2, 3, 4, 5, 6, 7],
-            color: "#4DA9CE",
-            active: "#F3FF15",
-          },
-          {
-            id: 3,
-            name: "Kanto",
-            prefs: [8, 9, 10, 11, 12, 13, 14],
-            color: "#9899C9",
-            active: "#F3FF15",
-          },
-          {
-            id: 4,
-            name: "Chubu & Hokuriku",
-            prefs: [15, 16, 17, 18, 19, 20, 21, 22, 23],
-            color: "#9AD47F",
-            active: "#F3FF15",
-          },
-          {
-            id: 5,
-            name: "Kinki",
-            prefs: [24, 25, 26, 27, 28, 29, 30],
-            color: "#B3DD53",
-            active: "#F3FF15",
-          },
-          {
-            id: 6,
-            name: "Chugoku",
-            prefs: [31, 32, 33, 34, 35],
-            color: "#ECDD28",
-            active: "#F3FF15",
-          },
-          {
-            id: 7,
-            name: "Shikoku",
-            prefs: [36, 37, 38, 39],
-            color: "#FDB42A",
-            active: "#F3FF15",
-          },
-          {
-            id: 8,
-            name: "Kyushu & Okinawa",
-            prefs: [40, 41, 42, 43, 44, 45, 46, 47],
-            color: "#F96F39",
-            active: "#F3FF15",
-          },
-        ],
-      });
+  //     window.svgJapan({
+  //       element: "#my-map-container",
+  //       uniformly: true,
+  //       uniformColor: "#00984F",
+  //       // width: "auto",
+  //       height: "calc(100vh - 120px)",
+  //       regionality: true,
+  //       type: "deform",
+  //       width: "100vw",
+  //       regions: [
+  //         {
+  //           id: 1,
+  //           name: "Hokkaido",
+  //           prefs: [1],
+  //           color: "#689ECA",
+  //           active: "#F3FF15",
+  //         },
+  //         {
+  //           id: 2,
+  //           name: "Tohoku",
+  //           prefs: [2, 3, 4, 5, 6, 7],
+  //           color: "#4DA9CE",
+  //           active: "#F3FF15",
+  //         },
+  //         {
+  //           id: 3,
+  //           name: "Kanto",
+  //           prefs: [8, 9, 10, 11, 12, 13, 14],
+  //           color: "#9899C9",
+  //           active: "#F3FF15",
+  //         },
+  //         {
+  //           id: 4,
+  //           name: "Chubu & Hokuriku",
+  //           prefs: [15, 16, 17, 18, 19, 20, 21, 22, 23],
+  //           color: "#9AD47F",
+  //           active: "#F3FF15",
+  //         },
+  //         {
+  //           id: 5,
+  //           name: "Kinki",
+  //           prefs: [24, 25, 26, 27, 28, 29, 30],
+  //           color: "#B3DD53",
+  //           active: "#F3FF15",
+  //         },
+  //         {
+  //           id: 6,
+  //           name: "Chugoku",
+  //           prefs: [31, 32, 33, 34, 35],
+  //           color: "#ECDD28",
+  //           active: "#F3FF15",
+  //         },
+  //         {
+  //           id: 7,
+  //           name: "Shikoku",
+  //           prefs: [36, 37, 38, 39],
+  //           color: "#FDB42A",
+  //           active: "#F3FF15",
+  //         },
+  //         {
+  //           id: 8,
+  //           name: "Kyushu & Okinawa",
+  //           prefs: [40, 41, 42, 43, 44, 45, 46, 47],
+  //           color: "#F96F39",
+  //           active: "#F3FF15",
+  //         },
+  //       ],
+  //     });
 
-      const allPaths = document.querySelectorAll(".prefecture-map");
+  //     const allPaths = document.querySelectorAll(".prefecture-map");
 
-      allPaths.forEach((path) => {
-        const jpName = path.getAttribute(
-          "data-name"
-        ) as keyof typeof engPrefNames;
-        const engName = engPrefNames[jpName];
-        path.setAttributeNS(null, "data-name", engName);
+  //     allPaths.forEach((path) => {
+  //       const jpName = path.getAttribute(
+  //         "data-name"
+  //       ) as keyof typeof engPrefNames;
+  //       const engName = engPrefNames[jpName];
+  //       path.setAttributeNS(null, "data-name", engName);
 
-        path.addEventListener("click", function () {
-          const trailPageURL = `trails/${engName.toLowerCase()}`;
-          router.push(trailPageURL);
-        });
-      });
-    }
-  }, []);
+  //       path.addEventListener("click", function () {
+  //         const trailPageURL = `trails/${engName.toLowerCase()}`;
+  //         router.push(trailPageURL);
+  //       });
+  //     });
+  //   }
+  // }, []);
 
   return (
     <Container maxWidth="lg" sx={{ mt: 10 }}>
@@ -176,9 +181,11 @@ const prefectures = () => {
               height={500}
             />
           </div>
+          <BrowserPrefList engPrefNames={engPrefNames} />
         </BrowserView>
-
-        <div id="my-map-container"></div>
+        <MobileView>
+          <MobilePrefList data={engPrefArr} />
+        </MobileView>
       </div>
     </Container>
   );
